+++ src/mo_frontend/typing.mli
@@ -21,7 +21,7 @@ val check_lib
   -> string option
   -> Syntax.lib
   -> scope Diag.result
-val check_actors : ?check_actors:bool -> scope -> Syntax.prog list -> unit Diag.result
+val check_actors : ?check_actors:bool -> scope -> Syntax.prog -> unit Diag.result
 
 val check_stab_sig : scope -> Syntax.stab_sig -> Type.stab_sig Diag.result